feat: set description and id for the VIP 310
